[main] 33 monitus crud - authorize with policy

+++ app/Http/Controllers/MonitorController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\MonitorStoreRequest;
 use App\Models\Monitor;
-use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class MonitorController extends Controller
@@ -23,15 +22,14 @@ class MonitorController extends Controller
 
     public function update(MonitorStoreRequest $request, Monitor $monitor)
     {
+        $this->authorize('update', $monitor);
         $monitor->update($request->validated());
         return to_route('home');
     }
 
     public function store(MonitorStoreRequest $request)
     {
-        if (Gate::denies('create-monitor')) {
-            abort(403, 'You are not allowed to create a monitor.');
-        }
+        $this->authorize('create', Monitor::class);
 
         Monitor::create($request->validated());
         return to_route('home');
@@ -39,6 +37,7 @@ class MonitorController extends Controller
 
     public function destroy(Monitor $monitor)
     {
+        $this->authorize('delete', $monitor);
         $monitor->delete();
         return to_route('home');
     }
